Move examples to tests
