--- examples/1d_classical_ising.jl
@@ -1,147 +1,0 @@
-using ITensors
-using ITensorsGrad
-using Zygote
-using LinearAlgebra
-using Test
-
-@testset "Zygote ITensors" begin
-  # Matrix
-  T(β) = [β ^ 2 β ^ 2
-          β ^ 2 β ^ 2]
-
-  # ITensor
-  function A(β, i = Index(2, "i"))
-    return itensor(T(β), i', dag(i))
-  end
-
-  β = 3.5
-
-  @show β
-
-  ##############################################
-  Z(β) = T(β)[1, 1]
-  Zit(β) = swapprime(A(β), 0 => 1)[1, 1]
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = (T(β) * T(β))[1, 1]
-  Zit(β) = (Aᵦ = A(β); (Aᵦ' * Aᵦ)[1, 1])
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ', 2 => 0) * Aᵦ)[])
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); tr(Aᵦ' * Aᵦ; plev = 0 => 2))
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ' * Aᵦ, 2 => 1) * dag(δ(inds(Aᵦ))))[])
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (Aᵦ' * Aᵦ * dag(δ(mapprime(inds(Aᵦ), 1 => 2))))[])
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = (T(β) + 2 * T(β) * T(β))[1, 1]
-  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ, 1 => 2) + 2 * Aᵦ' * Aᵦ)[1, 1])
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); tr(product(Aᵦ, Aᵦ)))
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = (2 * T(β) * T(β))[1, 1]
-  Zit(β) = (Aᵦ = A(β); (2 * Aᵦ' * Aᵦ)[1, 1])
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ'', 3 => 0) * Aᵦ' * Aᵦ)[])
-  @test Z(β) ≈ Zit(β)
-  @test_broken Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = T(β)[1, 1]
-  Zit(β) = A(β)'[1, 1]
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = T(β)[1, 1]
-  Zit(β) = A(β)[1, 1]
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = (T(β) * 3)[1, 1]
-  Zit(β) = (A(β) * 3)[1, 1]
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = (T(β) + 2 * T(β))[1, 1]
-  Zit(β) = (Aᵦ = A(β); (Aᵦ + 2 * Aᵦ)[1, 1])
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β))
-  Zit(β) = (Aᵦ = A(β); (Aᵦ * dag(δ(inds(Aᵦ))))[])
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (Aᵦ * Aᵦ)[])
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(2 * T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (2 * Aᵦ * Aᵦ)[])
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = (2 * T(β))[1, 1]
-  Zit(β) = (2 * A(β))[1, 1]
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = (T(β) + 2 * T(β))[1, 1]
-  Zit(β) = (Aᵦ = A(β); (Aᵦ + 2 * Aᵦ)[1, 1])
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (Aᵦ * dag(Aᵦ))[])
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-
-  ##############################################
-  Z(β) = tr(T(β) * T(β))
-  Zit(β) = (Aᵦ = A(β); (swapprime(Aᵦ, 0 => 1) * Aᵦ)[])
-  @test Z(β) ≈ Zit(β)
-  @test Z'(β) ≈ Zit'(β)
-end
-
