+++ src/ITensorsGrad.jl
@@ -5,16 +5,19 @@ using ITensors
 using NDTensors
 using LinearAlgebra
 
+# Use for truncated SVD
 #using BackwardsLinalg
+
+# Use for testing
 #using FiniteDifferences
-#using Zygote
 
 using ITensors: setinds
 
 import Base: +, adjoint
 import ChainRulesCore: rrule
-import ITensors: itensor
+import ITensors: itensor, dag, prime
 
+include("ITensors.jl")
 include("chainrules.jl")
 
 end
