+++ src/chainrules.jl
@@ -16,6 +16,7 @@ function rrule(::Type{<:Diag}, data)
   return Diag(data), Diag_pullback
 end
 
+# XXX Not sure about this definition
 function rrule(::Type{<:Tensor},
                is::IndexSet, st::TensorStorage)
   function Tensor_pullback(ΔΩ)
@@ -95,37 +96,11 @@ function rrule(::typeof(+), A::ITensor, B::ITensor)
 end
 
 # XXX TODO: This is needed for conj(::NamedTuple) error
+# Maybe instead define overload of conj(::NamedTuple)
 function rrule(::typeof(dag), T::ITensor)
   function dag_pullback(ȳ::ITensor)
     return (NO_FIELDS, ȳ)
   end
   return dag(T), dag_pullback
 end
-
-# To get around Zygote keyword argument issue
-import ITensors: dag
-function dag(T::ITensor)
-  TT = conj(tensor(T))
-  return itensor(store(TT), dag(inds(T)))
-end
-
-# TODO: are these needed?
-# TODO: define this as (A::Base.RefValue + B::ITensor) = A[] + B
-# and then define conversion of NamedTuple to ITensor
-(A::Base.RefValue + B::ITensor) =
-  itensor(A[].store) + B
-
-# TODO: are these needed?
-(A::ITensor + B::Base.RefValue) = B + A
-
-# TODO: is this needed?
-adjoint(A::Base.RefValue) = adjoint(A[])
-
-adjoint(A::NamedTuple{(:store, :inds), Tuple{ITensorT, Nothing}} where {ITensorT <: ITensor}) = prime(A)
-
-prime(A::NamedTuple{(:store, :inds), Tuple{ITensorT, Nothing}} where {ITensorT <: ITensor}) = prime(A.store)
-
-adjoint(A::NamedTuple{(:store, :inds), Tuple{T, Nothing}} where {T <: TensorStorage}) = A.store
-
-itensor(A::ITensor) = A
