+++ test/runtests.jl
@@ -1,31 +1,154 @@
-using ITensorsGrad
+using FiniteDifferences # For testing gradients
 using ITensors
+using ITensorsGrad
+using LinearAlgebra
 using Test
 using Zygote
 
-@testset "ITensorsGrad.jl" begin
+@testset "Zygote ITensors" begin
+  # Matrix
+  T(β) = [β ^ 2 β ^ 2
+          β ^ 2 β ^ 2]
 
-  mat(x) = [x*exp(x) x*exp(-x); x*exp(-x) x*exp(x)]
-  function iten(x, i)
-    return itensor(mat(x), i', i)
+  # ITensor
+  function A(β, i = Index(2, "i"))
+    return itensor(T(β), i', dag(i))
   end
 
-  @testset "Simple gradient" begin
+  β = 3.5
 
-    function f(x, i)
-      T = iten(x, i)
-      return T[i'=>1, i=>1]^2 + T[i'=>2, i=>1]^2
-    end
+  @show β
 
-    function g(x)
-      M = mat(x)
-      return M[1, 1]^2 + M[2, 1]^2
-    end
+  ##############################################
+  Z(β) = T(β)[1, 1]
+  Zit(β) = swapprime(A(β), 0 => 1)[1, 1]
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
 
-    i = Index(2)
-    x = 2.5
-    @test gradient(x -> f(x, i), x)[1] ≈ gradient(g, x)[1]
+  ##############################################
+  Z(β) = (T(β) * T(β))[1, 1]
+  Zit(β) = (Aᵦ = A(β); (Aᵦ' * Aᵦ)[1, 1])
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
 
-  end
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ', 2 => 0) * Aᵦ)[])
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
 
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); tr(Aᵦ' * Aᵦ; plev = 0 => 2))
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ' * Aᵦ, 2 => 1) * dag(δ(inds(Aᵦ))))[])
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (Aᵦ' * Aᵦ * dag(δ(mapprime(inds(Aᵦ), 1 => 2))))[])
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = (T(β) + 2 * T(β) * T(β))[1, 1]
+  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ, 1 => 2) + 2 * Aᵦ' * Aᵦ)[1, 1])
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); tr(product(Aᵦ, Aᵦ)))
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = (2 * T(β) * T(β))[1, 1]
+  Zit(β) = (Aᵦ = A(β); (2 * Aᵦ' * Aᵦ)[1, 1])
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β) * T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (mapprime(Aᵦ'', 3 => 0) * Aᵦ' * Aᵦ)[])
+  @test Z(β) ≈ Zit(β)
+  @test_broken Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = T(β)[1, 1]
+  Zit(β) = A(β)[1, 1]
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = T(β)[1, 1]
+  Zit(β) = A(β)'[1, 1]
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = T(β)[1, 1]^2 + T(β)[2, 1]^2
+  Zit(β) = (Aᵦ = A(β); (i′, i) = inds(Aᵦ); Aᵦ[i′=>1, i=>1]^2 + Aᵦ[i′=>2, i=>1]^2)
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = (T(β) * 3)[1, 1]
+  Zit(β) = (A(β) * 3)[1, 1]
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = (T(β) + 2 * T(β))[1, 1]
+  Zit(β) = (Aᵦ = A(β); (Aᵦ + 2 * Aᵦ)[1, 1])
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β))
+  Zit(β) = (Aᵦ = A(β); (Aᵦ * dag(δ(inds(Aᵦ))))[])
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (Aᵦ * Aᵦ)[])
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(2 * T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (2 * Aᵦ * Aᵦ)[])
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = (2 * T(β))[1, 1]
+  Zit(β) = (2 * A(β))[1, 1]
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = (T(β) + 2 * T(β))[1, 1]
+  Zit(β) = (Aᵦ = A(β); (Aᵦ + 2 * Aᵦ)[1, 1])
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (Aᵦ * dag(Aᵦ))[])
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
+
+  ##############################################
+  Z(β) = tr(T(β) * T(β))
+  Zit(β) = (Aᵦ = A(β); (swapprime(Aᵦ, 0 => 1) * Aᵦ)[])
+  @test Z(β) ≈ Zit(β)
+  @test Z'(β) ≈ Zit'(β)
 end
+
